==> 2_variable_selection.Rmd <==
---
title: "Variable selection"
output: html_notebook
---

```{r}
#automatic install of packages if they are not readily available
list.of.packages <- c(
  "tidyverse", 
  "ggplot2",
  "patchwork",
  "cowplot",
  "sf",
  "viridis",
  "kableExtra",
  "ggcorrplot",
  "sp",
  "spdep",
  "adespatial",
  "RSpectra",
  "Matrix",
  "parallel",
  "doParallel",
  "foreach",
  "factoextra",
  "ape",
  "remotes"
  )

new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]

#installing missing packages
if(length(new.packages) > 0){
  install.packages(new.packages, dep=TRUE)
}

#loading packages
for(package.i in list.of.packages){
  suppressPackageStartupMessages(library(package.i, character.only = TRUE))
}

#installing spatialRF
remotes::install_github(
  repo = "blasbenito/spatialRF", 
  ref = "main",
  force = TRUE
  )
library(spatialRF)

#loading functions
source("functions.R")

#removing scientific notation
options(scipen = 9999)

#setting ggplot2 theme
ggplot2::theme_set(theme_bw())

#loading data
load("ecoregions_plant_diversity.RData")

#removing columns with variance and mode zero from ecoregions
numeric.columns <-  sapply(ecoregions, is.numeric)
numeric.columns.var.0 <- names(
  which(
    round(
      apply(
        ecoregions[, numeric.columns], 
        2, 
        var
        ), 
      4
      ) == 0
    )
  )
numeric.columns.mode.0 <- names(
  which(
    apply(
      ecoregions[, numeric.columns], 
      2, 
      statistical_mode
    ) == 0
  )
)
#of these, only 5 are predictors
numeric.columns.mode.0 <- c(
  "human_footprint_minimum",
  "landcover_bare_percent_minimum",
  "landcover_herbs_percent_minimum",
  "landcover_trees_percent_minimum",
  "topography_elevation_minimum"
)
#other problematic columns
problematic.columns <- c(
  "landcover_bare_percent_maximum",
  "landcover_herbs_percent_maximum",
  "landcover_ndvi_minimum"
)

#removing columns
ecoregions <- ecoregions[, !(colnames(ecoregions) %in% c(numeric.columns.var.0, numeric.columns.mode.0, problematic.columns))]

rm(list.of.packages, new.packages, package.i)
```

# The data

## ecoregions

Here I focus on the world's ecoregions with an area equal or larger than 10.000 squared kilometers and that are at least at 1000 kilometers from any other ecoregion. These 664 ecoregions are represented once per row in the dataset `ecoregions`, that has 257 columns following this structure:

-   Columns relative to the ecoregion identification, named `ecoregion_x` describing the name, biome, realm, continent and area of the ecoregion in squared kilometers.
-   Columns with proxies of bias in the species data of the ecoregions, named `bias_x`, with the number of records, records per area unit, number of species per record, and number of species per record per squared kilometre.
-   Columns relative to plant species richness, named `richness_x`, with richness values for diferent taxonomic levels (from species to family) for different groups of plants (all plants, vascular plants, trees, and grasses).
-   Columns named `rarity_x` with *rarity-weighted richness* values (sum of 1 divided by the number of presences of each species present in the ecoregion, and sum of 1 divided by the number of ecoregions in which each species can be found) for different taxonomic groups (all plants, vascular plants, trees, and grasses) and different taxonomic levels (species and genus).
-   Columns named `exclusive_x` with the number and percentage of exclusive taxa of the given ecoregion for different plant groups and taxonomic levels.
-   Features of the ecoregion's neighbors (either direct neighbors, or other ecoregions within 100km when the ecoregion is isolated by the ocean), with columns named `neighbors_x`, containing values such as the number of neighbors, the sum of the neighbor's area, the total percentage of shared edge with those neighbors, their average aridity, and their number of species, genera, classes, and vamilies.
-   Columns describing the impact of human activity in the ecoregion, named `human_x`, including the conservation assessment by Dinerstein et al. (2017), the human footpring (Sanders citation), and the human population (SIEDAC citation).
-   Columns named `climate_x` describing the average climatic conditions within a given ecoregion.
-   Columns named `landcover_x` representing cover percentage of bare soil, herbs and trees (citation modis continuous fields) and NDVI (citation of modis product).
-   Geography columns named `geo_x` with the range and average of the latitude and longitude of each ecoregion.
-   Columns representing elevation stats, named `topography_elevation_x`.
-   A complete set of columns named `fragmentation_x` describing the fragmentation of each ecoregion computed at 5km resolution with the package `landscapemetrics` (citation).
-   Betadiversity measures (columns are named `betadiversity_x`) between each given ecoregion and its neighbors for different taxonomic levels and plant groups.

```{r, fig.width = 12, fig.height=6, echo = FALSE, fig.cap="Species richness of vascular plants by ecoregion"}
ggplot2::ggplot(data = ecoregions_polygons) +
    ggplot2::geom_sf(aes(fill = richness_species_vascular), size = 0.1) +
    ggplot2::scale_fill_viridis_c() + 
    ggplot2::labs(fill = "Richness species")
```
```{r, fig.width = 12, fig.height=6, echo = FALSE, fig.cap="Species richness of trees by ecoregion"}
ggplot2::ggplot(data = ecoregions_polygons) +
    ggplot2::geom_sf(aes(fill = richness_species_trees), size = 0.1) +
    ggplot2::scale_fill_viridis_c() + 
    ggplot2::labs(fill = "Tree richness")
```
```{r, fig.width = 12, fig.height=6, echo = FALSE, fig.cap="Species richness of grasses (Poaceae) by ecoregion"}
ggplot2::ggplot(data = ecoregions_polygons) +
    ggplot2::geom_sf(aes(fill = richness_species_grasses), size = 0.1) +
    ggplot2::scale_fill_viridis_c() + 
    ggplot2::labs(fill = "Grasses richness")
```

## ecoregions_pairs


Each ecoregion in `ecoregions` is paired with all the other ecoregions in the table `ecoregions_pairs`, that has 4440232 rows and 170 columns. This table is organized in several groups of columns. As a general rule, the features of a target ecoregion are named `ecoregion_x` while the features of its pair are named `neighbor_x`.

-   `ecoregion_name` and `neighbor_name`: names of the pair.
-   `ecoregion_area` and `neighbor_area`: area in squared kilometres of each ecoregion in the pair.
-   `ecoregion_climate_aridity_index_average` and `neighbor_climate_aridity_index_average`: the average aridity index of each ecoregion in the pair.
-   The `x_geo_x` columns contain the average latitude and longitude of each ecoregion pair, so basically these are the ecoregion centroids.
-   The columns `x_toporgraphy_elevation_average` contain the average elevation values of each ecoregion.
-   The columns `connection_x` are data on the connectivity between both ecoregions in the pair. Questions such as "is the pair of ecoregions in the same biome/continent?", or "are they direct neighbors?", and the length of the shared edge when they are neighbors are answered by the data in these columns.
-   The columns `richness_x_diff` have the differences in richness for each pair of ecoregions.
-   The columns `climate_x_diff`, `human_x_diff`, `landcover_x_diff`, and `topography_elevation_diff` represent differences in mean climatic, human activity, landcover, and elevation values between ecoregions pairs.
-   The columns `betadiversity_x` represent betadiversity values between each ecoregion pair for different taxonomic levels.
-   The overlap in the environmental space, or lack thereof, for each pair of ecoregions is stored in the columns `environmental_overlap_x`.

```{r, fig.width = 9, fig.height = 5, echo = FALSE, fig.cap="Subset of ecoregions_pairs centered on the Gibson desert"}
#subsetting the table
plot.df <- ecoregions_pairs %>% 
  dplyr::filter(
    ecoregion_name == "Gibson desert",
    neighbor_name %in% ecoregions[ecoregions$ecoregion_continent == "Oceania", "ecoregion_name"]
  )

#network plot
#plotting airport network
p1 <- ggplot2::ggplot(data = ecoregions_polygons[ecoregions_polygons$ecoregion_continent == "Oceania", ]) +
  ggplot2::geom_sf() +
  ggplot2::geom_segment(
    data = plot.df, aes(
    x = ecoregion_geo_longitude_average,
    xend = neighbor_geo_longitude_average,
    y = ecoregion_geo_latitude_average,
    yend = neighbor_geo_latitude_average,
    color = betadiversity_a_percent_families_vascular
  ),
  size = 1
  ) +
  ggplot2::scale_color_viridis_c(direction = -1) + 
  ggplot2::xlab("Longitude") +
  ggplot2::ylab("Latitude") +
  ggplot2::coord_sf(xlim = c(110, 175)) +
  ggplot2::labs(color = expression(paste("% shared families \n of vascular plants")))
print(p1)
rm(p1, plot.df)
```

# Assessing sampling bias in the plant richness data

There are several indicators of bias in the `ecoregions` dataset:

  + `bias_records_per_km2`: Number of GBIF records per squared kilometer. Gives an idea of the density of the spatial sampling of the ecoregion. A low number indicates that the ecoregion has not been sample thoroughly, or that the data available in GBIF Plantae might be missing data from such ecoregion. 
  + `bias_species_per_record`: Represents the number of species discovered per presence record available. A high number indicates that the ecoregion is likely poorly sampled. However, it also indicates that the sampling is diverse, that is, it is not focused in a particular group of species, which at the end might lead to a representative species pool in some cases.
  + `bias_species_per_record_per_km2`: Computed as `bias_species_per_record / bias_records_per_km2`. 
  
```{r, fig.width = 12, fig.height = 9, fig.cap="Bias columns in ecoregions", echo = FALSE}
p1 <- plot_variable_distribution(
  ecoregions_polygons,
  ecoregions,
  variable = "bias_records_per_km2",
  title = "Records/km2",
  viridis.direction = -1,
  binwidth = 0.5
  )

p2 <- plot_variable_distribution(
  ecoregions_polygons,
  ecoregions,
  variable = "bias_species_per_record",
  title = "Species/records",
  binwidth = 0.05
  )

p3 <- plot_variable_distribution(
  ecoregions_polygons,
  ecoregions,
  variable = "bias_species_per_record_per_km2",
  title = "Species/(records/km)",
  binwidth = 100
  )

p1 / p2 / p3 + plot_annotation(tag_levels = 'a')
rm(p1, p2, p3)
```

According to the `bias` columns, there is a small set of ecoregions that are *hyper-sampled*, especially in central europe (panels *a* and *b* in Figure 3). A small number of ecoregions, mostly drylands in the central Sahara and Eurasia, are close to one species per record, indicating that these ecoregions are *under-sampled*, and most likely plant taxa might be missing from their species lists (panels *c* and *d* in Figure 3). Finally, when assessing the overall bias through the species per record per kilometer we can see that most ecoregions have an adequate sampling, except for hyperarid regions of the Sahara, the Alashan plateau semi-desert, the East Siberian taiga, and a few others in the eastern shores of the Caspian Sea.

The bias variables `bias_records_per_km2` and `bias_species_per_record_per_km2` are log transformed.

```{r}
#computing log of the bias variables
ecoregions$bias_log_records_per_km2 <- log(ecoregions$bias_records_per_km2)
ecoregions$bias_log_species_per_record_per_km2 <- log(ecoregions$bias_species_per_record_per_km2)
ecoregions_polygons$bias_log_records_per_km2 <- log(ecoregions_polygons$bias_records_per_km2)
ecoregions_polygons$bias_log_species_per_record_per_km2 <- log(ecoregions_polygons$bias_species_per_record_per_km2)
```

To condense the information provided by the three bias variables into a single one, below I compute a PCA, assess the proportion of variance of the first factor, and extract it into `ecoregions` and `ecoregions_polygons` with the name `sampling_bias`.

```{r}
bias.pca <- prcomp(
  x = ecoregions[, c(
    "bias_log_records_per_km2",
    "bias_log_species_per_record_per_km2",
    "bias_species_per_record"
  )]
)
summary(bias.pca)$importance
```

The first PCA factor explains a proportion of variance of `r summary(bias.pca)$importance[2, 1]`, and therefore can be safely used as representative of the three bias variables.

```{r}
ecoregions$sampling_bias <- ecoregions_polygons$sampling_bias <- bias.pca$x[, 1]
```

Below I plot again each bias variable along with the new `sampling_bias` variable.

```{r, fig.width = 12, fig.height = 3, fig.cap="First PCA factor of the bias variables.", echo = FALSE}
plot_variable_distribution(
  ecoregions_polygons,
  ecoregions,
  variable = "sampling_bias",
  title = "First PCA factor of bias variables (sampling_bias)",
  viridis.direction = -1,
  binwidth = 1
  )
rm(bias.pca)
```

# Understanding the relationship between response variables

## ecoregions dataset

In the `ecoregions` dataset there are several groups of response variables computed for different taxonomic levels, and relationships among these variables are to be expected. Understanding these relationships will later help understand why models for different response variables show similar structures.

```{r, fig.width=9, fig.height=6, fig.cap="Correlogram between diversity variables for vascular species."}
ecoregions.response.variables <- ecoregions[, c(
  "richness_species_vascular",
  "rarity_weighted_richness_species_vascular",
  "mean_rarity_species_vascular",
  "exclusive_percent_species_vascular",
  "betadiversity_a_species_vascular",
  "betadiversity_a_percent_species_vascular",
  "betadiversity_Bsim_species_vascular",
  "betadiversity_R_species_vascular",
  "betadiversity_R_percent_species_vascular"
)] %>% 
  scale() %>% 
  as.data.frame()

ggcorrplot(
  cor(ecoregions.response.variables), 
  hc.order = TRUE, 
  type = "lower",
   lab = TRUE
  )
```

The correlogram shows that three pairs of response variables show high positive correlation:

  + `mean_rarity_species_vascular` and `exclusive_percent_species_vascular`.
  + `rarity_weighted_richness_species_vascular` and `betadiversity_a_species_vascular`.
  + `richness_species_vascular` and `rarity_weighted_richness_species_vascular`
  + `richness_species_vascular` and `betadiversity_a_species_vascular`.

One pair is highly negatively correlated:

  + `betadiversity_R_percent_species_vascular` and `betadiversity_a_percent_species_vascular`
  
```{r, fig.width=9, fig.height=6, fig.cap="Correlation between groups of response variables."}
p1 <- ggplot2::ggplot(data = ecoregions.response.variables) + 
  ggplot2::aes(
    x = mean_rarity_species_vascular,
    y = exclusive_percent_species_vascular
    ) + 
  ggplot2::geom_point(alpha = 0.3) +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::geom_abline(intercept = 0, slope = 1, col = "red4")

p2 <- ggplot2::ggplot(data = ecoregions.response.variables) + 
  ggplot2::aes(
    x = rarity_weighted_richness_species_vascular,
    y = betadiversity_a_species_vascular
    ) + 
  ggplot2::geom_point(alpha = 0.3) +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::geom_abline(intercept = 0, slope = 1, col = "red4")

p3 <- ggplot2::ggplot(data = ecoregions.response.variables) + 
  ggplot2::aes(
    x = richness_species_vascular,
    y = rarity_weighted_richness_species_vascular
    ) + 
  ggplot2::geom_point(alpha = 0.3) +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::geom_abline(intercept = 0, slope = 1, col = "red4")

p4 <- ggplot2::ggplot(data = ecoregions.response.variables) + 
  ggplot2::aes(
    x = richness_species_vascular,
    y = betadiversity_a_species_vascular
    ) + 
  ggplot2::geom_point(alpha = 0.3) +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::geom_abline(intercept = 0, slope = 1, col = "red4")

p5 <- ggplot2::ggplot(data = ecoregions.response.variables) + 
  ggplot2::aes(
    x = betadiversity_R_percent_species_vascular,
    y = betadiversity_a_percent_species_vascular
    ) + 
  ggplot2::geom_point(alpha = 0.3) +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::geom_abline(intercept = 0, slope = -1, col = "red4")
  
cowplot::plot_grid(p1, p2, p3, p4, p5, align = "hv")

rm(p1, p2, p3, p4, p5, ecoregions.response.variables)
```

Considering these correlations, only `betadiversity_a_x` and `betadiversity_R_percent_x` seem to perfectly mirror other variables, and therefore I remove them from the pool of response variables.

```{r}
#re-doing the betadiversity response variables
response.variables.betadiversity <- unique(
  c(
    colnames(ecoregions)[grepl(
      "^betadiversity_a_percent", 
      colnames(ecoregions)
    )],
    colnames(ecoregions)[grepl(
      "^betadiversity_Bsor", 
      colnames(ecoregions)
    )],
    colnames(ecoregions)[grepl(
      "^betadiversity_Bsim", 
      colnames(ecoregions)
    )],
    colnames(ecoregions)[grepl(
      "^betadiversity_R", 
      colnames(ecoregions)
    )][!(colnames(ecoregions)[grepl(
      "^betadiversity_R", 
      colnames(ecoregions)
    )] %in% colnames(ecoregions)[grepl(
      "^betadiversity_R_percent", 
      colnames(ecoregions)
    )]
    )
    ]
  )
)
```


## ecoregions_pairs dataset

Here I repeat the operation for the table `ecoregions_pairs`.

```{r, fig.width=9, fig.height=6, fig.cap="Correlogram between diversity variables for vascular species."}
ecoregions.pairs.response.variables <- ecoregions_pairs[, c(
  "diff_richness_species_vascular",
  "diff_richness_genera_vascular",
  "diff_richness_families_vascular",
  "diff_richness_classes_vascular",
  "betadiversity_a_species_vascular",
  "betadiversity_a_percent_species_vascular",
  "betadiversity_Bsim_species_vascular",
  "betadiversity_R_species_vascular",
  "betadiversity_R_percent_species_vascular"
)] %>% 
  scale() %>% 
  as.data.frame()

ggcorrplot(
  cor(ecoregions.pairs.response.variables), 
  hc.order = TRUE, 
  type = "lower",
   lab = TRUE
  )

rm(ecoregions.pairs.response.variables)
```
The correlogram shows that only close taxonomic levels are highly correlated, and therefore no variables are removed at this stage.


# Reducing multicollinearity in the predictors

## ecoregions

There is a large number of redundant predictors in the `ecoregions` dataset. To reduce multicollinearity here I use the functions `spatialRF::auto_vif()` and `spatialRF::auto_cor()`. Both allow the user to define a preference order in the selection of variables, a maximum correlation threshold (`cor.threshold`) and a maximum variance inflation factor threshold (`vif.threshold`). Variables listed first in the `preference.order` argument take preference over the other variables during the elimination of variables performed by the correlation and variance inflation factor analyses.

Below I define the preference order of selection by prioritizing sampling bias and ecoregion area, then climate variables, the data about the neighbors of each ecoregion, the geographic fragmentation of the ecoregion, the average NDVI, topography, and human impact on the ecoregion.

```{r}
#gathering the names of every predictor
predictive.variables <- c(
  "sampling_bias", 
  "ecoregion_area_km2",
  colnames(ecoregions)[grepl("^neighbors", colnames(ecoregions))],
  colnames(ecoregions)[grepl("^human", colnames(ecoregions))],
  colnames(ecoregions)[grepl("^climate", colnames(ecoregions))],
  colnames(ecoregions)[grepl("^landcover", colnames(ecoregions))],
  colnames(ecoregions)[grepl("^topography", colnames(ecoregions))],
  colnames(ecoregions)[grepl("^fragmentation", colnames(ecoregions))]
)

#defining a preference order
preference.order <- c(
  "sampling_bias",
  "ecoregion_area_km2",
  "climate_aridity_index_average",
  "climate_hypervolume",
  "climate_velocity_lgm_average",
  "climate_bio1_average",
  "climate_bio12_average",
  "climate_bio15_average",
  "neighbors_count",
  "neighbors_area",
  "neighbors_percent_shared_edge",
  "fragmentation_cohesion",
  "fragmentation_division",
  "landcover_ndvi_average",
  "topography_elevation_range",
  "topography_elevation_average",
  "human_population",
  "human_footprint_average"
)

#check that all variables in preference.order are in predictive.variables
if(sum(preference.order %in% predictive.variables) != length(preference.order)){
  warning("There are variables in preference.order not in predictive.variables")
}

#multicollinearity reduction
ecoregions.selected.variables <- spatialRF::auto_cor(
  x = ecoregions[, predictive.variables],
  preference.order = preference.order,
  cor.threshold = 0.5,
  verbose = FALSE
) %>% 
  spatialRF::auto_vif(
    preference.order = preference.order,
    vif.threshold = 5,
    verbose = FALSE
  )

rm(predictive.variables, preference.order)
```

The multicollinearity analysis returns a list of variable names that are not collinear with the others.

```{r, echo = FALSE}
kableExtra::kbl(
  data.frame(`Selected predictors` = ecoregions.selected.variables$selected.variables)
) %>% kable_styling(latex_options = c("HOLD_position", "striped"))
```

## ecoregions_pairs

```{r}
predictive.variables <- c(
  colnames(ecoregions_pairs)[grepl("^connection", colnames(ecoregions_pairs))],
  colnames(ecoregions_pairs)[grepl("^diff_human", colnames(ecoregions_pairs))],
  colnames(ecoregions_pairs)[grepl("^diff_climate", colnames(ecoregions_pairs))],
  colnames(ecoregions_pairs)[grepl("^diff_landcover", colnames(ecoregions_pairs))],
  colnames(ecoregions_pairs)[grepl("^diff_topography", colnames(ecoregions_pairs))],
  colnames(ecoregions_pairs)[grepl("^environmental", colnames(ecoregions_pairs))]
)

#defining a preference order
preference.order <- c(
  "environmental_overlap_sorensen",
  "diff_climate_aridity_index",
  "diff_climate_velocity_lgm_average",
  "diff_climate_bio1_average", 
  "diff_climate_bio12_average",
  "connection_distance",
  "connection_are_neighbors",
  "connection_same_continent",
  "connection_shared_edge",
  "connection_area_diff",
  "diff_human_population",
  "diff_human_footprint_average",
  "diff_human_population_density",
  "diff_topography_elevation" 
)

#check that all variables in preference.order are in predictive.variables
if(sum(preference.order %in% predictive.variables) != length(preference.order)){
  warning("There are variables in preference.order not in predictive.variables")
}

#multicollinearity reduction
ecoregions.pairs.selected.variables <- spatialRF::auto_cor(
  x = ecoregions_pairs[, predictive.variables],
  preference.order = preference.order,
  cor.threshold = 0.5,
  verbose = FALSE
) %>% 
  spatialRF::auto_vif(
    preference.order = preference.order,
    vif.threshold = 5,
    verbose = FALSE
  )

rm(predictive.variables, preference.order)
```

```{r, echo = FALSE}
kableExtra::kbl(
  data.frame(`Selected predictors` = ecoregions.pairs.selected.variables$selected.variables)
) %>% kable_styling(latex_options = c("HOLD_position", "striped"))
```
# Betadiversity and distance matrices from ecoregions_pairs

From the dataset `ecoregions_pairs` several betadiversity and distance matrices can be extracted. Here I compute the matrices of the percentage of shared taxa and Sorensen's Beta for species, genera, and families, and the matrix of geographic distances.

```{r}
#object to store results
distance_matrices <- list()

#variables to iterate over
variables <- c(
  "betadiversity_a_percent_species_vascular",
  "betadiversity_a_percent_genera_vascular",
  "betadiversity_a_percent_families_vascular",
  "betadiversity_Bsor_species_vascular",
  "betadiversity_Bsor_genera_vascular",
  "betadiversity_Bsor_families_vascular",
  "connection_distance"
)

#iterating over variables
for(variable.i in variables){
  
  #distance
  d.i <- reshape2::acast(
    ecoregions_pairs,
    ecoregion_name~neighbor_name,
    value.var = variable.i
  )
  
  #diagonal to zero
  diag(d.i) <- 0
  
  #ordering like ecoregions and saving
  distance_matrices[[variable.i]] <- d.i[ecoregions$ecoregion_name, ecoregions$ecoregion_name]
  
}

#rounding to reduce size
distance_matrices$betadiversity_a_percent_species_vascular <- round(distance_matrices$betadiversity_a_percent_species_vascular, 1)

distance_matrices$betadiversity_a_percent_genera_vascular <- round(distance_matrices$betadiversity_a_percent_genera_vascular, 1)

distance_matrices$betadiversity_a_percent_families_vascular <- round(distance_matrices$betadiversity_a_percent_families_vascular, 1)

distance_matrices$betadiversity_Bsor_species_vascular <- round(distance_matrices$betadiversity_Bsor_species_vascular, 3)

distance_matrices$betadiversity_Bsor_genera_vascular <- round(distance_matrices$betadiversity_Bsor_genera_vascular, 3)

distance_matrices$betadiversity_Bsor_families_vascular <- round(distance_matrices$betadiversity_Bsor_families_vascular, 3)

distance_matrices$connection_distance <- round(distance_matrices$connection_distance, 0)

rm(d.i, variables, variable.i)
```

# Organizing the data in lists

To avoid cluttering the R environment with objects, here I organize the datasets `ecoregions` and `ecoregions_pairs` in two lists with their respective names, that also store information about the names of the response variables and predictors of each conceptual group.

```{r, echo = FALSE}
#organizing the ecoregions data

#copy of ecoregions
ecoregions.data <- ecoregions

#as a list
ecoregions <- list()

#adding dataframe
ecoregions$data <- ecoregions.data

#column names
ecoregions$names <- colnames(ecoregions$data)

#names of response variables
ecoregions$response.variables$richness <- ecoregions$names[grepl("^richness", ecoregions$names)]
ecoregions$response.variables$rarity.weighted.richness <- ecoregions$names[grepl("^rarity_weighted_richness", ecoregions$names)]
ecoregions$response.variables$mean.rarity <- ecoregions$names[grepl("^mean_rarity", ecoregions$names)]
ecoregions$response.variables$exclusivity <- ecoregions$names[grepl("^exclusive", ecoregions$names)]
ecoregions$response.variables$betadiversity <- response.variables.betadiversity
ecoregions$response.variables$all <- c(
  ecoregions$response.variables$richness,
  ecoregions$response.variables$rarity.weighted.richness,
  ecoregions$response.variables$mean.rarity,
  ecoregions$response.variables$exclusivity,
  ecoregions$response.variables$betadiversity
)

#names of predictors
ecoregions$predictor.variable.names <- ecoregions.selected.variables$selected.variables

#adding distance matrix
ecoregions$distance.matrix <- distance_matrices$connection_distance

#adding centroids
sf::st_geometry(ecoregions_centroids) <- "geom_centroids"
ecoregions$centroids <- ecoregions_centroids
ecoregions$polygons <- ecoregions_polygons

rm(ecoregions_centroids, ecoregions_polygons, ecoregions.data, response.variables.betadiversity)
```

```{r}
#organizing the ecoregions_pairs data
ecoregions_pairs.data <- ecoregions_pairs

#creating list
ecoregions_pairs <- list()

#adding the data
ecoregions_pairs$data <- ecoregions_pairs.data

#column names
ecoregions_pairs$names <- colnames(ecoregions_pairs$data)

#names of response variables
ecoregions_pairs$response.variables$richness <- ecoregions_pairs$names[grepl("^richness", ecoregions_pairs$names)]
ecoregions_pairs$response.variables$betadiversity <- unique(c(
  ecoregions$names[grepl("^betadiversity_a", ecoregions$names)],
  ecoregions$names[grepl("^betadiversity_a_percent", ecoregions$names)],
  ecoregions_pairs$names[grepl("^betadiversity_Bsor", ecoregions_pairs$names)],
  ecoregions_pairs$names[grepl("^betadiversity_Bsim", ecoregions_pairs$names)],
  ecoregions_pairs$names[grepl("^betadiversity_R", ecoregions_pairs$names)],
  ecoregions_pairs$names[grepl("^betadiversity_R_percent", ecoregions_pairs$names)]
  ))
ecoregions_pairs$response.variables$all <- c(
  ecoregions_pairs$response.variables$richness,
  ecoregions_pairs$response.variables$betadiversity
)

#selecting predictors
ecoregions_pairs$predictor.variable.names <- ecoregions.pairs.selected.variables$selected.variables

rm(ecoregions_pairs.data)
```

To separate arid and non-arid ecoregions I use the threshold 0.65 in the variable "climate_aridity_index_average". I select the indices in `ecoregions$data` to select the cases above and below the threshold and compose the datasets `ecoregions_arid` and `ecoregions_humid`.

```{r}
indices.arid <- which(ecoregions$data$climate_aridity_index_average >= 0.65)
indices.humid <- which(ecoregions$data$climate_aridity_index_average < 0.65)
```

From these indices I can prepare the lists `ecoregions_arid` and `ecoregions_humid` as follows.

```{r}
#copies of ecoregions
ecoregions_arid <- ecoregions_humid <- ecoregions

#subsetting objects in ecoregions_arid
ecoregions_arid$data <- ecoregions_arid$data[indices.arid, ]
ecoregions_arid$distance.matrix <- ecoregions_arid$distance.matrix[indices.arid, indices.arid]
ecoregions_arid$centroids <- ecoregions_arid$centroids[indices.arid, ]
ecoregions_arid$polygons <- ecoregions_arid$polygons[indices.arid, ]

#subsetting objects in ecoregions_humid
ecoregions_humid$data <- ecoregions_humid$data[indices.humid, ]
ecoregions_humid$distance.matrix <- ecoregions_humid$distance.matrix[indices.humid, indices.humid]
ecoregions_humid$centroids <- ecoregions_humid$centroids[indices.humid, ]
ecoregions_humid$polygons <- ecoregions_humid$polygons[indices.humid, ]
```


```{r}
save(ecoregions, ecoregions_arid, ecoregions_humid, ecoregions_pairs, distance_matrices, file = "ecoregions_plant_diversity_ready.RData")
```
