==> 3_analysis.Rmd <==
---
title: "Analysis"
output: html_notebook
---

```{r, echo = FALSE}
#automatic install of packages if they are not readily available
list.of.packages <- c(
  "tidyverse", 
  "ggplot2",
  "patchwork",
  "sf",
  "viridis",
  "kableExtra",
  "sp",
  "spdep",
  "adespatial",
  "parallel",
  "doParallel",
  "foreach",
  "lavaan",
  "ranger",
  "quantable",
  "factoextra"
  )

new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]

#installing missing packages
if(length(new.packages) > 0){
  install.packages(new.packages, dep=TRUE)
}

#loading packages
for(package.i in list.of.packages){
  suppressPackageStartupMessages(library(package.i, character.only = TRUE))
}

#installing spatialRF
remotes::install_github(
  repo = "blasbenito/spatialRF", 
  ref = "main",
  force = TRUE
  )
library(spatialRF)

#loading functions
source("functions.R")

#removing scientific notation
options(scipen = 9999)

#setting ggplot2 theme
ggplot2::theme_set(theme_bw())

#loading data
load("ecoregions_plant_diversity_ready.RData")

rm(list.of.packages, new.packages, package.i)
```


## Drivers of richness, rarity, and betadiversity in the world's ecoregions

Here I use the `ecoregions`, `ecoregions_arid`, and `ecoregions_humid` datasets to assess how climate, humans, fragmentation, topography, and landcover shape diversity measures such as richness, mean rarity, percentage of exclusive species, and betadiversity with neighboring ecoregions. Considering the different ranges and distributions of the range variables, to obtain comparable outcomes and simplify the modeling process I will be fitting spatial regression models with Random Forest using the [`spatialRF` package](https://github.com/BlasBenito/spatialRF).

There are 65 response variables divided in tree groups of diversity measures (richness, rarity, and betadiversity) with nine different measures for three taxonomic levels (species, genera, families) of three target groups (vascular plants and two of its subgroups, trees and grasses). The only exception for these combinations are grasses, that belong to the *Poaceae* family, and therefore do not have "families" metrics. To facilitate the organization of results here I generate a template dataframe with the response variable name, the group, the taxonomic level, and the target group.

```{r}
#defining the different groups
groups <- c(
  "richness",
  "rarity",
  "betadiversity"
)

measures <- list(
  richness = "richness",
  rarity = c(
    "rarity_weighted_richness",
    "exclusive",
    "exclusive_percent",
    "mean_rarity"
  ),
  betadiversity = c(
    "betadiversity_a_percent",
    "betadiversity_Bsor",
    "betadiversity_Bsim",
    "betadiversity_R"
  )
)

taxonomic.groups <- c(
  "species",
  "genera",
  "families"
)

functional.groups <- c(
  "vascular",
  "trees",
  "grasses"
)

predictors <- c(
  ecoregions$predictor.variable.names,
  "spatial_predictors (max)",
  "spatial_predictors (mean)",
  "spatial_predictors (median)",
  "spatial_predictors (min)"
  )

datasets <- c(
  "ecoregions", 
  "ecoregions_arid", 
  "ecoregions_humid"
  )

#precomputing size
n.rows <- length(ecoregions$response.variables$all) * length(predictors) * length(datasets)

#empty dataframe
output.template <- data.frame(
  response_variable = rep(NA, n.rows),
  group = rep(NA, n.rows),
  measure = rep(NA, n.rows),
  taxonomy = rep(NA, n.rows),
  functional = rep(NA, n.rows),
  dataset = rep(NA, n.rows),
  predictor = rep(NA, n.rows)
)

#filling dataframe
row.i <- 1

for(group.i in groups){
  for(measure.i in measures[[group.i]]){
    for(taxonomy.i in taxonomic.groups){
      for(functional.i in functional.groups){
        for(dataset.i in datasets){
          for(predictor.i in predictors){
        
        #generating response variable name
        response.variable.name.i <- paste(
          measure.i, 
          taxonomy.i, 
          functional.i,
          sep = "_"
        )
        
        #if the response variable name is not a column of ecoregions$data, next
        if(!(response.variable.name.i %in% ecoregions$names)){
          next
        }
        
        #filling the dataframe
        output.template[row.i, "group"] <- group.i
        output.template[row.i, "measure"] <- measure.i
        output.template[row.i, "taxonomy"] <- taxonomy.i
        output.template[row.i, "functional"] <- functional.i
        output.template[row.i, "response_variable"] <- response.variable.name.i
        output.template[row.i, "dataset"] <- dataset.i
        output.template[row.i, "predictor"] <- predictor.i
        
        #one more line
        row.i <- row.i + 1
        
          }
        }
      }
    }
  }
}

#removing empty rows
output.template <- na.omit(output.template)

#adding to ecoregions$response.variables$all
ecoregions$response.variables$all <- output.template$response_variable

rm(group.i, measure.i, taxonomy.i, functional.i, response.variable.name.i, row.i, groups, measures, taxonomic.groups, functional.groups, predictors, datasets, n.rows)
```
